OOP tovabbi ervenyesitese, pydantic modellek kulon lettek mozgatva.
